requirements.txt
tensorflow~=2.11.0
numpy~=1.24.2
pillow~=9.4.0